# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "contact": {
            "name": "Krisna Cipta Raksa",
            "email": "kciptaraksa@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/article": {
            "get": {
                "tags": [
                    "Get All Article"
                ],
                "summary": "Fetch article data (using MongoDB)",
                "description": "Fetch all data article.",
                "operationId": "fetcharticle",
                "responses": {
                    "200": {
                        "description": "Fetch data success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "response_while_fetch_data_article_success": {
                                        "summary": "Response while fetch data successfully",
                                        "value": {
                                            "status": "success",
                                            "message": "Data fetched successfully.",
                                            "data": "Article data"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to fetch data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "response_while_fetch_data_article_failed": {
                                        "summary": "Response while fetch data failed",
                                        "value": {
                                            "status": "error",
                                            "message": "Failed to fetch data",
                                            "data": ""
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Request without Token",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "example": "Unauthorized."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/article/store": {
            "post": {
                "tags": [
                    "Store Article"
                ],
                "summary": "Store article data (using MongoDB)",
                "description": "Store article data.",
                "operationId": "storearticle",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {}
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Title 1",
                                    "category": "Hukum",
                                    "content": "Content title 1",
                                    "tags": [
                                        "hukum",
                                        "pidana"
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "response_while_store_data_article_success": {
                                        "summary": "Response while store data successfully",
                                        "value": {
                                            "status": "success",
                                            "message": "Article created successfully.",
                                            "data": "Article data inserted"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to store data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "response_while_store_data_article_failed": {
                                        "summary": "Response while store data failed",
                                        "value": {
                                            "status": "error",
                                            "message": "Article created failed.",
                                            "data": ""
                                        }
                                    },
                                    "response_while_store_data_article_failed_because_mandatory_field_empty": {
                                        "summary": "Response while store data failed",
                                        "value": {
                                            "status": "error",
                                            "message": "Mandatory field cannot be empty",
                                            "data": "Validation Information"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Request without Token",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "example": "Unauthorized."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/article/{id}/show": {
            "get": {
                "tags": [
                    "Get Article By ID"
                ],
                "summary": "Fetch article data by ID (using MongoDB)",
                "description": "Fetch article data by ID.",
                "operationId": "fetcharticlebyid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fetch data success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "response_while_fetch_data_article_byid_success": {
                                        "summary": "Response while fetch data successfully",
                                        "value": {
                                            "status": "success",
                                            "message": "Data fetched successfully.",
                                            "data": "Article data"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to fetch data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "response_while_fetch_data_article_byid_failed": {
                                        "summary": "Response while fetch data failed",
                                        "value": {
                                            "status": "error",
                                            "message": "Failed to fetch data",
                                            "data": ""
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Request without Token",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "example": "Unauthorized."
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Failed to fetch data because ID not given",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "response_while_fetch_data_article_byid_failed_because_no_id": {
                                        "summary": "Response while fetch data failed because ID not given",
                                        "value": {
                                            "status": "error",
                                            "message": "Page Not Found",
                                            "data": ""
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get authentication token",
                "name": "apiAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}